refactor(DUP): device disconnection
Moves the handling of device disconnection in the `Core.Device` module
